About page complete
Eric if you're reading this, what's left is the following:

Resources page, this one I recommend looking into collapsible/accordion from bootstrap which will give you the open-close UI

We also need to clean up the login and signuppage.

Also there is a public speaking and stories page. Neither of these are built but shouldn't take long to hack together.