Finished restructuring client-side code.
We are now using webpack for all locally develop JS assets. This is because turbolink/sprockets is being difficult and not letting us declare templates after the fact.

Follow the guideline in mount.js and also import the correct dependencies and webpack will take care of the rest.

Next task: add more to the control-component stylings